backend/requirements.txt
flask==3.1.2
flask-cors==6.0.1
google-generativeai==0.8.5
python-dotenv==1.0.1
nltk==3.9.1
unidecode==1.3.8
PyPDF2==3.0.1